nosql/ana_filter/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pandas==2.2.3
psycopg2-binary==2.9.9
numpy==1.26.4
tensorflow==2.17.0
